+++ src/backend/db/sqlite/table.py
