+++ src/backend/db/vector/redis_client.py
