+++ src/backend/knowledge_base/document/doc_embedder.py
