+++ src/backend/knowledge_base/image/__tagger_cleaner.py
