+++ src/backend/knowledge_base/image/__tagger_model.py
