+++ src/backend/knowledge_base/image/image_embedder.py
