+++ src/backend/knowledge_base/image/image_tagger.py
