+++ src/backend/lib_manager.py
