+++ src/backend/library/image/image_lib.py
