+++ src/backend/library/image/image_lib_table.py
