+++ src/backend/library/image/image_lib_vector_db.py
