+++ src/backend/library/lib_item.py
