+++ src/backend/library/scan_record_tracker.py
