+++ src/backend/utils/exceptions/lib_errors.py
